% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkExperiment.R
\name{benchmarkExperiment}
\alias{benchmarkExperiment}
\title{Benchmark different models on the training data set using cross-validation
Input: lrns(list of learners), task, list of measures 
Output: lrns(hyperparameter tuned list of learners-(fnr,tpr,f1,auc) by default}
\usage{
benchmarkExperiment(lrns, task, measures = list(fnr, tpr, f1, auc))
}
\description{
Benchmark different models on the training data set using cross-validation
Input: lrns(list of learners), task, list of measures 
Output: lrns(hyperparameter tuned list of learners-(fnr,tpr,f1,auc) by default
}
