% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestLearnerIndex.R
\name{bestLearnerIndex}
\alias{bestLearnerIndex}
\title{Find the best model in the benchmark result}
\usage{
bestLearnerIndex(bmr)
}
\description{
Find the best model in the benchmark result
}
