% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_select.R
\name{feature_select}
\alias{feature_select}
\title{Feature selection function}
\usage{
feature_select(X_train, y_train, threshold = 0.05)
}
\arguments{
\item{X_train}{The training set of the data}

\item{y_train}{The target of the data}

\item{threshold}{user input threshold used for stopping criteria}
}
\value{
List of best features
}
\description{
Performs forward selection of features in the data by starting with an empty model,
and iteratively adds features that improve the model's score. The algorithm stops once
the increase of the accuracy from an additional features is smaller than the threshold.
}
\examples{
feature_select(X_train, y_train, threshold=0.05)
}
