% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_data.R
\name{miss_data}
\alias{miss_data}
\title{Handle missing data in a data frame.}
\usage{
miss_data(x_train, x_test, strategy)
}
\arguments{
\item{x_train}{The train set dataframe with missing values in it.}

\item{x_test}{The test set DataFrame with missing values in it.}

\item{strategy}{The imputation strategy as a string}
}
\value{
x_train and x_test dataframe without any missing values.
}
\description{
Impute missing values with the strategy (mean, median) selected by the user.
}
\examples{
x_train <- data.frame('x' = c(400, NA, 330, NA), 'y' = c(24, NA, 30, 560))
x_test <- data.frame('x' = c(NA, 130, 240, NA), 'y' = c(NA, 300, 450, 100))
miss_data(x_train, x_test, 'mean')
}
