% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramSet.R
\name{paramSet}
\alias{paramSet}
\title{Search space for Hyperparameters
Currently supports major learning algorithms}
\usage{
paramSet(lrn)
}
\description{
Search space for Hyperparameters
Currently supports major learning algorithms
}
