% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{Evaluate performance of prediction algorithm
input: pred object, list of measures for performance analysis}
\usage{
performance(pred, measures = list(ppv, tpr, f1))
}
\description{
Evaluate performance of prediction algorithm
input: pred object, list of measures for performance analysis
}
