% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{Used for prediction using the trained model
input: threshold, trained model object(mod), test_task (containing test data) 
output: pred object}
\usage{
prediction(mod, test_task, threshold = 0.5)
}
\description{
Used for prediction using the trained model
input: threshold, trained model object(mod), test_task (containing test data) 
output: pred object
}
