% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectLearner.R
\name{selectLearner}
\alias{selectLearner}
\title{User chooses the learner to be used for training by looking at the output of analyseThresholdVsPerformance() function}
\usage{
selectLearner(lrns = NULL, name = NULL)
}
\description{
User chooses the learner to be used for training by looking at the output of analyseThresholdVsPerformance() function
}
