% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Used to train the model
input: lrn, task
output: trained model( mod)}
\usage{
train(lrn, task)
}
\description{
Used to train the model
input: lrn, task
output: trained model( mod)
}
