% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatClassImbalance.R
\name{treatClassImbalance}
\alias{treatClassImbalance}
\title{Treat Class Imbalance. Default rate is 8}
\usage{
treatClassImbalance(task, rate = 8)
}
\description{
Treat Class Imbalance. Default rate is 8
}
