% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneLearners.R
\name{tuneLearners}
\alias{tuneLearners}
\title{Automatic Hyperparameter tuning using cross-validation of a list of learners
Input: lrns(list of learners)
Output: lrns(hyperparameter tuned list of learners}
\usage{
tuneLearners(lrns, task)
}
\description{
Automatic Hyperparameter tuning using cross-validation of a list of learners
Input: lrns(list of learners)
Output: lrns(hyperparameter tuned list of learners
}
