% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_fun.R
\name{baseline_fun}
\alias{baseline_fun}
\title{Baseline function}
\usage{
baseline_fun(X, y, type = "regression")
}
\arguments{
\item{X}{The features of the training set}

\item{y}{The target of the training set}

\item{type}{What kind of supervised machine learning to use, regression or classification}
}
\value{
The list of the result of 5-fold cross validation for the chosen model.
}
\description{
Gives the scoring metrics of LinearRegression or LogisticRegression.
}
\examples{
baseline_fun(X, y, type = 'regression')
}
