% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_select.R
\name{feature_select}
\alias{feature_select}
\title{Feature selection function}
\usage{
feature_select(X_train, y_train, threshold = 0.05)
}
\arguments{
\item{X_train}{The training set of the data}

\item{y_train}{The target of the training set data}

\item{threshold}{user input threshold used for bounding features selected}
}
\value{
List of best features
}
\description{
Performs forward selection of features in the data by starting with an empty model,
and iteratively adds features that improve the model's score. The algorithm stops once
the decrease of the accuracy from an additional features is bigger than the threshold.
}
\examples{
library(MASS)
attach(Boston)
data <- Boston
X <- data [,0:13]
y <- data [,14]
feature_select(X, y, threshold=0.05)
}
