% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr2.R
\name{preProcess}
\alias{preProcess}
\title{Function for preprocessing data
turn character into factors}
\usage{
preProcess(dataset)
}
\description{
Function for preprocessing data
turn character into factors
}
